refactor: move scoring mode config to BaseConfig, driven by registry scoringModes
