+++ packages/web/src/features/questions/index.ts
@@ -1,4 +1,8 @@
-import type { QuestionOptions, QuestionType } from "@razzia/common/types/game"
+import type {
+  QuestionOptions,
+  QuestionType,
+  ScoringMode,
+} from "@razzia/common/types/game"
 import * as multi from "@razzia/web/features/questions/multi"
 import * as single from "@razzia/web/features/questions/single"
 import type {
@@ -10,6 +14,7 @@ import type { ComponentType } from "react"
 interface QuestionRegistryEntry {
   labelKey: string
   defaultOptions?: QuestionOptions
+  scoringModes?: ScoringMode[]
   AnswerComponent: ComponentType<AnswerComponentProps>
   ConfigComponent: ComponentType
   SolutionPicker: ComponentType<SolutionPickerProps>
