+++ packages/web/src/features/questions/multi/index.ts
@@ -1,5 +1,8 @@
 import { SCORING_MODES } from "@razzia/common/constants"
-import type { MultiQuestionOptions } from "@razzia/common/types/game"
+import type {
+  MultiQuestionOptions,
+  ScoringMode,
+} from "@razzia/common/types/game"
 
 export { default as AnswerComponent } from "@razzia/web/features/questions/multi/components/MultiAnswers"
 
@@ -12,3 +15,9 @@ export const labelKey = "quizz:questionType.multi"
 export const defaultOptions: MultiQuestionOptions = {
   scoringMode: SCORING_MODES.BALANCED,
 }
+
+export const scoringModes: ScoringMode[] = [
+  SCORING_MODES.STRICT,
+  SCORING_MODES.BALANCED,
+  SCORING_MODES.LENIENT,
+]
