+++ packages/web/src/locales/it/quizz.json
@@ -15,7 +15,6 @@
       "questionDisplayHint": "Durata prima che appaiano le risposte",
       "timings": "Tempi",
       "answerMode": "Modalità risposta",
-      "multiSelect": "Selezione multipla",
       "scoringMode": "Modalità punteggio",
       "scoringMode.strict": "Rigoroso",
       "scoringMode.balanced": "Bilanciato",
